[CI] [98608,98607,98605,98604,98603,98601,98600,98599,98598,98596,98595,98594,98593,98592,98590,98589,98588,98587,98586,98585]  <security> <feat> sync secureboot_xip changes

+++ cp/components/coredump/bk_coredump_mem.c
@@ -36,8 +36,8 @@ static bool bk_coredump_window_needs_ap_power(uint32_t addr, uint32_t size)
         return true;
     }
     /* PSRAM / QSPI media memory (AP / media power domain). */
-    if ((addr < (SOC_QSPI1_DATA_BASE + SOC_QSPI1_DATA_SIZE)) &&
-        (end > SOC_PSRAM0_DATA_BASE)) {
+    if ((addr < ((uint32_t)SOC_QSPI1_DATA_BASE + (uint32_t)SOC_QSPI1_DATA_SIZE)) &&
+        (end > (uint32_t)SOC_PSRAM0_DATA_BASE)) {
         return true;
     }
     return false;
